requirements.txt
streamlit
python-dotenv
langchain
langchain-community
PyPDF2
faiss-cpu
sentence-transformers
ollama